+++ src/components/ship/body/index.tsx
