+++ src/components/ship/exhaust/index.tsx
