+++ src/components/ship/physics/motions/yaw/yaw.test.tsx
